backend/requirements.txt
# Flask and web framework dependencies
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-CORS==5.0.0
Flask-Marshmallow==1.2.1

# Database dependencies
psycopg2-binary==2.9.9
SQLAlchemy==2.0.35

# Authentication and security
bcrypt==4.2.1
PyJWT==2.10.1

# API documentation
flask-restx==1.3.0
flasgger==0.9.7.1

# Testing dependencies
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0
pytest-mock==3.14.0

# Development dependencies
python-dotenv==1.0.1
marshmallow==3.23.2
marshmallow-sqlalchemy==1.1.0

# Performance testing
locust==2.35.0

# Utilities
click==8.1.7
Werkzeug==3.1.3